module reorganizing; not many real changes
- split models.py into {gardens, pipelines, steps}.py
- renamed garden.py -> client.py
- added /app subdirectory for CLI

`GardenScopes` class for backend auth

NOTE: this changes the GH action to skip integration tests; they're
not especially meaningful until the backend has been configured
to use the correct env variables

includes incidental fixes to requests_to_curl util

updating mocked auth bits in tests
