+++ garden_ai/__init__.py
@@ -1,4 +1,7 @@
-from .garden import GardenClient
-from .models import Garden, Pipeline, Step, step
+from .datacite import DataciteSchema
+from .client import GardenClient
+from .gardens import Garden
+from .pipelines import Pipeline
+from .steps import Step, step
 
-__all__ = ["GardenClient", "Garden", "Pipeline", "Step", "step"]
+__all__ = ["GardenClient", "Garden", "Pipeline", "Step", "step", "DataciteSchema"]
