+++ garden_ai/__main__.py
@@ -1,0 +1,3 @@
+from .app.garden import app
+
+app()
