+++ garden_ai/app/create.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+# module for the "garden create" subcommand
+# ("create" is also a typer app in its own right; we just have to register as a subcommand)
