+++ pyproject.toml
@@ -11,6 +11,7 @@ python = "^3.7"
 requests = "^2.20.0"
 globus-sdk = "^3.12.0"
 pydantic = "^1.10.2"
+typer = {extras = ["all"], version = "^0.7.0"}
 
 [tool.poetry.group.test]
 optional=true
